move client and healthcheck to elasticsearch directory
