+++ elasticsearch/client.go
